+++ config/config.go
@@ -162,11 +162,24 @@ type CommonTracing struct {
 	SamplingRate float64 `mapstructure:"samplingRate"`
 }
 
+type CommonPostgreSQL struct {
+	Host            string `mapstructure:"host"`
+	Port            int    `mapstructure:"port"`
+	User            string `mapstructure:"user"`
+	Password        string `mapstructure:"password"`
+	DBName          string `mapstructure:"dbname"`
+	SSLMode         string `mapstructure:"sslmode"`
+	MaxIdleConns    int    `mapstructure:"maxIdleConns"`
+	MaxOpenConns    int    `mapstructure:"maxOpenConns"`
+	ConnMaxLifetime int    `mapstructure:"connMaxLifetime"`
+}
+
 type Common struct {
-	CommonEtcd     CommonEtcd     `mapstructure:"common-etcd"`
-	CommonRedis    CommonRedis    `mapstructure:"common-redis"`
-	CommonRabbitMQ CommonRabbitMQ `mapstructure:"common-rabbitmq"`
-	CommonTracing  CommonTracing  `mapstructure:"common-tracing"`
+	CommonEtcd       CommonEtcd       `mapstructure:"common-etcd"`
+	CommonRedis      CommonRedis      `mapstructure:"common-redis"`
+	CommonRabbitMQ   CommonRabbitMQ   `mapstructure:"common-rabbitmq"`
+	CommonTracing    CommonTracing    `mapstructure:"common-tracing"`
+	CommonPostgreSQL CommonPostgreSQL `mapstructure:"common-postgresql"`
 }
 
 type ConnectBase struct {